requirements.txt
django==4.2
djangorestframework==3.15.1
pymysql==1.1.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
